+++ tests/Providers/ResponseBuilderServiceProvider.php
@@ -44,5 +44,8 @@ class ResponseBuilderServiceProvider extends \MarcinOrlowski\ResponseBuilder\Res
 
 		$this->loadTranslationsFrom(__DIR__ . '/../../src/lang', 'response-builder');
 	}
+<<<<<<< HEAD
 
+=======
+>>>>>>> dev
 }